feat: Update ImageUploaderModal to use default density when initializing state
This commit modifies the ImageUploaderModal component to use the default density value when initializing the state for the density variable. Previously, the default density was set to "ascii", but it has been changed to an empty string. This change ensures that the component starts with the correct default value for the density, improving the user experience.